third_party/WebKit/LayoutTests/http/tests/inspector/network/x-frame-options-deny.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script>
function loadIFrameWithDownload()
{
    var iframe = document.createElement("iframe");
    iframe.setAttribute("src", "http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-deny.cgi");
    document.body.appendChild(iframe);
}

function test() {
  TestRunner.addSniffer(SDK.NetworkDispatcher.prototype, 'responseReceived', responseReceived);
  TestRunner.addSniffer(SDK.NetworkDispatcher.prototype, 'loadingFailed', loadingFailed);
  TestRunner.addSniffer(SDK.NetworkDispatcher.prototype, 'loadingFinished', loadingFinished);
  TestRunner.evaluateInPage('loadIFrameWithDownload()');

  function responseReceived(requestId, time, resourceType, response) {
    var request = TestRunner.networkLog.requestByManagerAndId(TestRunner.networkManager, requestId);
    if (/x-frame-options-deny\.cgi/.exec(request.url())) {
      TestRunner.addResult('Received response for x-frame-options-deny.cgi');
      TestRunner.addResult('SUCCESS');
      TestRunner.completeTest();
    }
  }

  function loadingFinished(requestId, finishTime) {
    var request = TestRunner.networkLog.requestByManagerAndId(TestRunner.networkManager, requestId);
    if (/x-frame-options-deny\.cgi/.exec(request.url()))
      TestRunner.completeTest();
  }

  function loadingFailed(requestId, time, localizedDescription, canceled) {
    var request = TestRunner.networkLog.requestByManagerAndId(TestRunner.networkManager, requestId);
    if (/x-frame-options-deny\.cgi/.exec(request.url())) {
      TestRunner.addResult('TODO(mkwst): This started failing when we moved XFO to the browser.');
      TestRunner.completeTest();
    }
  }
}
</script>
</head>
<body onload="runTest()">
<p>Tests that responseReceived is called on NetworkDispatcher for resource requests denied due to X-Frame-Options header.</p>
</body>
</html>
